+++ core/src/main/scala/swaydb/core/log/counter/CounterLog.scala
@@ -43,7 +43,7 @@ private[swaydb] object CounterLog {
   def persistent(dir: Path,
                  mmap: MMAP.Log,
                  mod: Long,
-                 fileSize: Long)(implicit bufferCleaner: ByteBufferSweeperActor,
+                 fileSize: Int)(implicit bufferCleaner: ByteBufferSweeperActor,
                                  forceSaveApplier: ForceSaveApplier,
                                  writer: LogEntryWriter[LogEntry.Put[Slice[Byte], Slice[Byte]]],
                                  reader: LogEntryReader[LogEntry[Slice[Byte], Slice[Byte]]]): IO[swaydb.Error.Log, PersistentCounterLog] =
